[IntersectionObserver] Fix the concept of "tracking document"
As explained in the bug, "tracking document" should be associated with
IntersectionObservation, not IntersectionObserver.

One side effect of this change is that for a given observer, each of its
observations runs its algorithm independently, so that notifications may
not be generated in the order in which the observations were created
(i.e., the order of calls to observer.observe()). To preserve the
idiomatic web platform behavior that notifications are delivered in
the order in which the targets were observed, this patch moves the
storage of pending notifications from IntersectionObserver to
IntersectionObservation.

BUG=879798
R=chrishtr@chromium.org

Change-Id: I9b6f6ad8a26387f0c072ccdff7f18cea9a88004c