[LayoutNG] Add LayoutBox::PercentageResolutionLogicalHeight.
This function allows LayoutNG to correctly query the percentage
resolution block_size in quirks mode. (see test case for an example of
where this happened).

We call this as we are crossing a legacy->LayoutNG boundary while
creating the constraint space from a LayoutObject.

The function will also be used in the CSS Layout API to provide the
correct percentage resolution size.

Cq-Include-Trybots: luci.chromium.try:linux_layout_tests_layout_ng
Change-Id: I295145996f2460de1b25f85ff3784df6d6edbd30