Change detune min/max limits
The max and min valures for the OscillatorNode.detune parameter has been
reduced according the latest version of the spec.  Make it so.

Also implement clamping of the total frequency to Nyquist so that we don't
generate infinity or NaN doing the computations.  This is ok because anything
beyond Nyquist would output 0 anyway.  But we're careful to preserve the slope
of automations; test added for this as well.

Update expectations for AudioParam/audioparam-nominal-range.html because of
the new warnings produced by clamping detune.

Bug: 865552, 860347
Test: the-oscillatornode-interface/detune-limiting.html
Change-Id: I35a214b0f3156ad8912718dcb277f1967f70df8f